Apply nits from the review
