+++ fuel-block-producer/src/ports.rs
@@ -1,8 +1,8 @@
-use anyhow::Result;
 use async_trait::async_trait;
 use fuel_core_interfaces::model::{
     ArcPoolTx,
     BlockHeight,
+    DaBlockHeight,
 };
 
 #[async_trait]
@@ -13,5 +13,11 @@ pub trait TxPool: Sync + Send {
         block_height: BlockHeight,
         // The upper limit for the total amount of gas of these txs
         max_gas: u64,
-    ) -> Result<Vec<ArcPoolTx>>;
+    ) -> anyhow::Result<Vec<ArcPoolTx>>;
+}
+
+#[async_trait::async_trait]
+pub trait Relayer: Sync + Send {
+    /// Get the best finalized height from the DA layer
+    async fn get_best_finalized_da_height(&self) -> anyhow::Result<DaBlockHeight>;
 }
