+++ fuel-poa-coordinator/src/service.rs
@@ -3,6 +3,7 @@ use crate::{
         DeadlineClock,
         OnConflict,
     },
+    ports::BlockProducer,
     Config,
     Trigger,
 };
@@ -13,11 +14,7 @@ use anyhow::{
 use fuel_core_interfaces::{
     block_importer::ImportBlockBroadcast,
     common::{
-        fuel_tx::{
-            Receipt,
-            Transaction,
-            UniqueIdentifier,
-        },
+        fuel_tx::UniqueIdentifier,
         prelude::{
             Signature,
             Word,
@@ -58,24 +55,6 @@ use tracing::{
     error,
     warn,
 };
-
-#[async_trait::async_trait]
-pub trait BlockProducer: Send + Sync {
-    // TODO: Right now production and execution of the block is one step, but in the future,
-    //  `produce_block` should only produce a block without affecting the blockchain state.
-    async fn produce_and_execute_block(
-        &self,
-        height: BlockHeight,
-        max_gas: Word,
-    ) -> anyhow::Result<ExecutionResult>;
-
-    async fn dry_run(
-        &self,
-        transaction: Transaction,
-        height: Option<BlockHeight>,
-        utxo_validation: Option<bool>,
-    ) -> anyhow::Result<Vec<Receipt>>;
-}
 
 pub struct RunningService {
     join: JoinHandle<()>,
